requirements.txt
# Fabric==1.12.0
Fabric3==1.12.0.post1
